pyproject.toml
[tool.poetry]
name = "newsletter-agent"
version = "0.0.1"
description = ""
authors = ["Or Levi <orlevi128@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.12"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
tenacity = "^9.0.0"
pydantic-ai = "^0.0.15"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
mypy = "^1.14.0"
ruff = "^0.8.4"
ipdb = "^0.13.13"
types-requests = "^2.32.0.20241016"
types-beautifulsoup4 = "^4.12.0.20241020"
python-dotenv = "^1.0.1"
types-pyyaml = "^6.0.12.20241221"

[tool.poe.tasks]
format = "ruff format src"
lint = "ruff check src"
test = "pytest"
mypy = "mypy -p newsletter_agent"
# ----
config_schema.script = "newsletter_agent.config_schema:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
